+++ crates/polars-core/src/series/amortized_iter.rs
